- fixed some gsize/guint arguments of gwy_file_get_contents()
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code@3630 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ branches/GWYDDION-1/gwyddion/modules/file/nt-mdt.c
@@ -363,7 +363,7 @@ static GwyContainer*
 mdt_load(const gchar *filename)
 {
     guchar *buffer;
-    guint size;
+    gsize size;
     GError *err = NULL;
     GwyDataField *dfield = NULL;
     GwyContainer *data = NULL;
